// src/types/CourseMapConfig.type.ts
type CourseMapConfig = {
  ratioX: number;
  ratioY: number;
  yard: number;
  tmStartX: number;
  tmStartY: number;
  tmEndX: number;
  tmEndY: number;
  imgStartX: number;
  imgStartY: number;
  cos: number;
  sin: number;
  scalarRatio: number;
  originalWidth: number;
  originalHeight: number;
  renderedWidth: number;
  renderedHeight: number;
};

export default CourseMapConfig;